% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{dt_group_by}
\alias{dt_group_by}
\title{group by columns and return a summarized table}
\usage{
dt_group_by(dt, group_by_cols, summarize_at, operation)
}
\arguments{
\item{dt}{input data.table}

\item{group_by_cols}{group by columns}

\item{summarize_at}{column summarize at}

\item{operation}{calculation operation, value should be one of following: sum, mean, median, max, min}
}
\value{
a summarized table
}
\description{
group by columns and return a summarized table
}
\examples{
data("dt_groups")
dataMojo::dt_group_by(dt_groups,
                      group_by_cols = c("group1", "group2"),
                      summarize_at = "A1",
                      operation = "mean")
}
