% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{filter_all_at}
\alias{filter_all_at}
\title{Filter all rows that meet requirements with selected columns}
\usage{
filter_all_at(dt, operator, cutoff_value, selected_cols)
}
\arguments{
\item{dt}{input data table}

\item{operator}{operator should be one of l, or g. l means less than, g means greater than}

\item{cutoff_value}{cutoff value}

\item{selected_cols}{selected columns from input data table}
}
\value{
filtered data table
}
\description{
Filter all rows that meet requirements with selected columns
}
\examples{
data("dt_values")
dataMojo::filter_all_at(dt_values, operator = "l", .1, c("A1", "A2"))
}
