% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{filter_any}
\alias{filter_any}
\title{Filter any rows that meeting requirements}
\usage{
filter_any(dt, operator, cutoff_value)
}
\arguments{
\item{dt}{input data.table}

\item{operator}{operator should be one of l, g. l means less than, g means greater than.}

\item{cutoff_value}{threshold value}
}
\value{
fitlered data table
}
\description{
Filter any rows that meeting requirements
}
\examples{
data("dt_values")
dataMojo::filter_any(dt_values, operator = "l", .1)
}
