% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{get_row_group_by}
\alias{get_row_group_by}
\title{Fetch one row from each grouped by group}
\usage{
get_row_group_by(dt, group_by_cols, fetch_row)
}
\arguments{
\item{dt}{input data table}

\item{group_by_cols}{group by columns}

\item{fetch_row}{first means to fetch first row and last means to fetch last row}
}
\value{
grouped by data table
}
\description{
Fetch one row from each grouped by group
}
\examples{
data("dt_groups")
dataMojo::get_row_group_by(dt_groups,
                           group_by_cols = c("group1", "group2"),
                           fetch_row = "first")
}
