% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{reshape_wider}
\alias{reshape_wider}
\title{Reshape data frame to a wider format}
\usage{
reshape_wider(dt, keep_cols, col_label, col_value)
}
\arguments{
\item{dt}{input data table}

\item{keep_cols}{columns to be kept}

\item{col_label}{columns that each unique values will be reshaped as a column name}

\item{col_value}{columns that fill the reshaped columns}
}
\value{
reshaped widen data table
}
\description{
Reshape data frame to a wider format
}
\examples{
data("dt_long")
dataMojo::reshape_wider(dt_long,
  keep_cols = c("Full_name"),
  col_label = c("Date_Type"),
  col_value = "Exact_date")

}
