% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{row_expand_dates}
\alias{row_expand_dates}
\title{Expand row given start and end dates}
\usage{
row_expand_dates(dt, start_date_col, end_date_col, new_name)
}
\arguments{
\item{dt}{input data table}

\item{start_date_col}{start date column}

\item{end_date_col}{end date column}

\item{new_name}{new generated column name}
}
\value{
expanded data table
}
\description{
Expand row given start and end dates
}
\examples{
dt_dates_simple <- data.table::data.table(
  Start_Date = as.Date(c("2020-02-03", "2020-03-01") ),
  End_Date = as.Date(c("2020-02-05", "2020-03-02") ),
  group = c("A", "B")
)
row_expand_dates(dt_dates_simple, "Start_Date", "End_Date", "Date")[]
}
