% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_str_reshape.R
\name{row_expand_pattern}
\alias{row_expand_pattern}
\title{Expand row based on pattern}
\usage{
row_expand_pattern(dt, col_name, split_by_pattern, new_name)
}
\arguments{
\item{dt}{input data table}

\item{col_name}{column to be expanded}

\item{split_by_pattern}{split based on pattern}

\item{new_name}{new generated column name}
}
\value{
expanded data table
}
\description{
Expand row based on pattern
}
\examples{
data("starwars_simple")
row_expand_pattern(starwars_simple, "films", ", ", "film")[]
}
