% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-starwars_simple.R
\docType{data}
\name{starwars_simple}
\alias{starwars_simple}
\title{starwars data}
\format{
a data table as example
}
\usage{
data(starwars_simple)
}
\description{
starwars data
}
\examples{
data(starwars_simple)
}
\author{
Jiena Gu McLellan, 2020-05-26
}
\keyword{datasets}
